+++ src/main/java/kingwant/hjjp/controller/RoleController.java
@@ -8,7 +8,6 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 import com.baomidou.mybatisplus.mapper.EntityWrapper;
-
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
 import kingwant.hjjp.base.PublicResult;
@@ -43,35 +42,29 @@ public class RoleController {
 	
 	@PostMapping("/addRole")
 	@ApiOperation(value = "添加角色", notes = "所需参数：name(名字);des(描述)")
-	public PublicResult<Map<String, Object>> addUser(@RequestBody kingwant.hjjp.entity.Role role) {
+	public PublicResult<Map<String, Object>> addRole(@RequestBody kingwant.hjjp.entity.Role role) {
 		
         if (KwHelper.isNullOrEmpty(role.getName())) {
-        	Map<String, Object> map=new HashMap<>();
-        	map.put("角色名不能为空！！", false);
-            return new PublicResult<>(PublicResultConstant.MiSSING_KEY_PARAMETERS_ERROR, map);
+            return new PublicResult<>(PublicResultConstant.MiSSING_KEY_PARAMETERS_ERROR, null);
         }
         role.setId(KwHelper.newID());
         roleMapper.insert(role);
-        Map<String, Object> map=new HashMap<>();
-        map.put("成功", true);
-        return new PublicResult<>(PublicResultConstant.SUCCESS, map);		
+        return new PublicResult<>(PublicResultConstant.SUCCESS, null);		
     }
-	
 	
 	@DeleteMapping("/delRoleById")
 	@ApiOperation(value = "删除角色", notes = "所需参数：id(角色id)")
-	public PublicResult<Map<String, Object>> delUserById(String id) {
-		Map<String, Object> map=new HashMap<>();
+	public PublicResult<Map<String, Object>> delRoleById(String id) {
+		if (ComUtil.isEmpty(id)  ) {
+            return new PublicResult<>(PublicResultConstant.MiSSING_KEY_PARAMETERS_ERROR, null);
+        }
 		try {
 			roleMapper.deleteById(id);
-			map.put("删除成功！！", true);
 		} catch (Exception e) {
-			map.put("删除失败！！", false);
-			return new PublicResult<>(PublicResultConstant.SQL_EXCEPTION, map);
+			return new PublicResult<>(PublicResultConstant.SQL_EXCEPTION, null);
 			// TODO: handle exception
 		}
-        return new PublicResult<>(PublicResultConstant.SUCCESS, map);
-		
+        return new PublicResult<>(PublicResultConstant.SUCCESS, null);
     }
 	
 	@GetMapping("/findById")
@@ -80,13 +73,11 @@ public class RoleController {
 		//String name = requestJson.getString("name");
 		//参数校验
         if (ComUtil.isEmpty(id)  ) {
-        	Map<String, Object> map=new HashMap<>();
-        	map.put("缺少关键参数", false);
-            return new PublicResult<>(PublicResultConstant.MiSSING_KEY_PARAMETERS_ERROR, map);
+            return new PublicResult<>(PublicResultConstant.MiSSING_KEY_PARAMETERS_ERROR, null);
         }        
 		kingwant.hjjp.entity.Role role=roleMapper.selectById(id);        
         Map<String, Object> map=new HashMap<>();
-        map.put("成功", role);
+        map.put("data", role);
         return new PublicResult<>(PublicResultConstant.SUCCESS, map);	
     }
 	
@@ -100,7 +91,7 @@ public class RoleController {
 		ew.like(!KwHelper.isNullOrEmpty(role.getDes()), "des", role.getDes());
 		List<kingwant.hjjp.entity.Role> list = roleMapper.selectList(ew);        
         Map<String, Object> map=new HashMap<>();
-        map.put("成功", list);
+        map.put("data", list);
         return new PublicResult<>(PublicResultConstant.SUCCESS, map);
     }
 
@@ -108,22 +99,16 @@ public class RoleController {
 	@ApiOperation(value = "修改角色", notes = "所需参数：id(必要，其他的为选填);name(名字);des(描述)")
 	public PublicResult<Map<String, Object>> updateRole(@RequestBody kingwant.hjjp.entity.Role role) {
 		User user4update=new User();
-		Map<String, Object> map=new HashMap<>();
 		if(KwHelper.isNullOrEmpty(role.getId())) {
-			Object message=false;
-			map.put("用户信息id不能为空", false);
-			return new PublicResult<>(PublicResultConstant.MiSSING_KEY_PARAMETERS_ERROR, map);
+			return new PublicResult<>(PublicResultConstant.MiSSING_KEY_PARAMETERS_ERROR, null);
 		}
 		try {
 			roleMapper.updateById(role);
 		} catch (Exception e) {
-			map.put("操作数据失败", false);
-			return new PublicResult<>(PublicResultConstant.SQL_EXCEPTION, map);
+			return new PublicResult<>(PublicResultConstant.SQL_EXCEPTION, null);
 			// TODO: handle exception
 		}
-
-        map.put("成功", true);
-        return new PublicResult<>(PublicResultConstant.SUCCESS, map);
+        return new PublicResult<>(PublicResultConstant.SUCCESS, null);
     }
